1.0.1
新增购物车模块

+++ reposition/modal_del.py
@@ -1,9 +1,10 @@
 from . import models
 
 
-def query_del(tb_name, nid):
-    try:
-        tb_name.objects.filter(id=nid).delete()
-        return True
-    except Exception as e:
-        return e
+def query_del(req, tb_name, nid):
+    if req.method == 'GET':
+        try:
+            tb_name.objects.filter(id=nid).delete()
+            return '成功'
+        except Exception as e:
+            return e
